requirements.txt
distro==1.9.0
et_xmlfile==2.0.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tabula-py==2.10.0
tzdata==2025.1
